プロフィール画面修正
