+++ app/views/posts/_bookmark_button.html.erb
@@ -1,5 +1,5 @@
 <% if current_user.bookmark?(post) %>
-  <%= render 'unbookmark', { post: post } %>
+  <%= render 'posts/unbookmark', { post: post } %>
 <% else %>
-  <%= render 'bookmark', { post: post } %>
+  <%= render 'posts/bookmark', { post: post } %>
 <% end %>
