+++ app/views/posts/show.html.erb
@@ -8,13 +8,7 @@
         <div class="card-body">
           <div class='row'>
             <div class='col-md-12'>
-              <% if @post.movie? %>
-                <% if @post.movie.file.content_type.include?('image/') %>
-                  <%= image_tag @post.movie.url, class: 'product-img' %>
-                <% else %>
-                  <%= video_tag @post.movie.url, class: 'product-img', controls: true, autobuffer: true %>
-                <% end %>
-              <% end %>
+              <%= video_tag @post.movie.url, class: 'product-img', controls: true, autobuffer: true %>
             </div>
             <div class='col-md-9'>
               <h3 style='display: inline;'><%= @post.title %></h3>
