--- app/views/profiles/_crud_menus.html.erb
@@ -1,12 +1,0 @@
-<ul class='crud-menu-btn list-inline float-right'>
-  <li class="list-inline-item">
-    <%= link_to edit_post_path(post), id: "button-edit-#{post.id}" do %>
-      <i class="bi bi-pencil-fill"></i>
-    <% end %>
-  </li>
-  <li class="list-inline-item">
-    <%= link_to post_path(post), id: "button-delete-#{post.id}", method: :delete, data: { confirm: t('defaults.message.delete_confirm') } do %>
-      <i class="bi bi-trash-fill"></i>
-    <% end %>
-  </li>
-</ul>
