+++ app/views/users/edit.html.erb
@@ -3,7 +3,7 @@
   <div class="row">
     <div class="col-md-10 offset-md-1">
       <h1><%= t('.title') %></h1>
-      <%= form_with model: @user, url: profile_path, local: true do |f| %>
+      <%= form_with model: @user, url: user_path, local: true do |f| %>
         <%= render "shared/error_messages", object: f.object %>
           <div class="form-group mb-3">
             <div class='mt-3 mb-3'>
