+++ config/routes.rb
@@ -1,10 +1,6 @@
 Rails.application.routes.draw do
-  get 'password_resets/new'
-  get 'password_resets/create'
-  get 'password_resets/edit'
-  get 'password_resets/update'
   root "static_pages#top"
-  resources :users, only: %i[new create]
+  resources :users
   resources :posts do
     resources :comments, only: %i[create destroy], shallow: true
     collection do
@@ -12,7 +8,6 @@ Rails.application.routes.draw do
     end
   end
   resources :bookmarks, only: %i[create destroy]
-  resource :profile, only: %i[show edit update]
   resources :password_resets, only: %i[new create edit update]
 
   mount LetterOpenerWeb::Engine, at: "/letter_opener" if Rails.env.development?
